pyproject.toml
[tool.poetry]
name = "bpi-fr-algo-credit-scoring"
version = "0.1.0"
description = "Algo for credit scoring"
authors = ["Thibaut Charré <thibaut.charre@gnomondigital.com>"]
readme = "README.md"
packages = [
    {include = "bpi_fr_algo_credit_scoring", from = "src"}
    ]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.3"
numpy = "1.23.2"
seaborn = "^0.12.2"
scikit-learn = "^1.3.0"
liac-arff = "^2.5.0"
sklearn = "^0.0.post5"
matplotlib = "^3.7.2"
xgboost = "^1.7.6"
datetime = "^5.1"
ipykernel = "^6.24.0"
llvmlite = "^0.40.1"
shap = "^0.42.0"
dash = "^2.11.1"
plotly-express = "^0.4.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
flake8 = "^6.0.0"
black = "^23.3.0"
pre-commit = "^3.3.3"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
